finished first implementation of grid search for dnn
